+++ core/ui/App.ts
@@ -1,0 +1,5 @@
+import {Game} from "../game";
+
+export function App(game: Game) {
+    return "";
+}
