+++ core/world.ts
@@ -1,7 +1,9 @@
 import {Animate} from "./components/com_animate.js";
 import {AudioSource} from "./components/com_audio_source.js";
 import {Camera} from "./components/com_camera.js";
+import {Draw} from "./components/com_draw.js";
 import {Light} from "./components/com_light.js";
+import {Move} from "./components/com_move.js";
 import {Render} from "./components/com_render.js";
 import {Transform} from "./components/com_transform.js";
 
@@ -9,7 +11,9 @@ const enum Component {
     Animate,
     AudioSource,
     Camera,
+    Draw,
     Light,
+    Move,
     Render,
     Transform,
 }
@@ -18,7 +22,9 @@ export const enum Has {
     Animate = 1 << Component.Animate,
     AudioSource = 1 << Component.AudioSource,
     Camera = 1 << Component.Camera,
+    Draw = 1 << Component.Draw,
     Light = 1 << Component.Light,
+    Move = 1 << Component.Move,
     Render = 1 << Component.Render,
     Transform = 1 << Component.Transform,
 }
@@ -30,8 +36,10 @@ export interface World {
     // Component data
     Animate: Array<Animate>;
     AudioSource: Array<AudioSource>;
+    Draw: Array<Draw>;
     Camera: Array<Camera>;
     Light: Array<Light>;
+    Move: Array<Move>;
     Render: Array<Render>;
     Transform: Array<Transform>;
 }
